Implement simplified physics calculations

+++ Assets/Scripts/Time/IFrameProcessor.cs
@@ -1,5 +1,5 @@
 ﻿namespace Time {
     public interface IFrameProcessor {
-        public void ProcessFrame(float frameTime);
+        public void ProcessFrame(double frameTime);
     }
 }
